Renamed folders and baseURL
