+++ ecorise-react-app/package.json
@@ -6,8 +6,8 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-icons": "^5.5.0",
-    "react-router-dom": "^7.6.3",
-    "react-scripts": "^5.0.1",
+    "react-router": "^6.30.1",
+    "react-router-dom": "^6.30.1",
     "web-vitals": "^5.0.3"
   },
   "scripts": {
@@ -27,5 +27,11 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "@testing-library/jest-dom": "^6.6.4",
+    "@testing-library/react": "^16.3.0",
+    "jest": "^27.5.1",
+    "react-scripts": "^5.0.1"
   }
 }
