+++ ecorise-react-app/src/Pickup/index.css
