+++ ecorise-react-app/src/Pickup/index.test.js
